requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langgraph>=0.2.60
langchain-core>=0.3.28
pydantic>=2.10.0
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0